Move Bundle convenience extension into its own file
