+++ Sources/MuxUploadSDK/Extensions/Bundle+Reporting.swift
@@ -1,0 +1,15 @@
+//
+//  Bundle+Reporting.swift
+//  
+
+import Foundation
+
+extension Bundle {
+    var appName: String? {
+        return object(forInfoDictionaryKey: "CFBundleName") as? String
+    }
+
+    var appVersion: String? {
+        return object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String
+    }
+}
