+++ Sources/MuxUploadSDK/InternalUtilities/UploadEvent.swift
@@ -27,13 +27,3 @@ struct UploadEvent: Codable {
 
     var regionCode: String?
 }
-
-extension Bundle {
-    var appName: String? {
-        return object(forInfoDictionaryKey: "CFBundleName") as? String
-    }
-
-    var appVersion: String? {
-        return object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String
-    }
-}
